+++ packages/switch/lion-switch-button.js
@@ -1,0 +1,3 @@
+import { LionSwitchButton } from './src/LionSwitchButton.js';
+
+customElements.define('lion-switch-button', LionSwitchButton);