+++ packages/switch/lion-switch.js
@@ -1,0 +1,3 @@
+import { LionSwitch } from './src/LionSwitch.js';
+
+customElements.define('lion-switch', LionSwitch);