+++ packages/button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.33](https://github.com/ing-bank/lion/compare/@lion/button@0.3.32...@lion/button@0.3.33) (2019-11-06)
+
+**Note:** Version bump only for package @lion/button
+
+
+
+
+
+## [0.3.32](https://github.com/ing-bank/lion/compare/@lion/button@0.3.31...@lion/button@0.3.32) (2019-11-01)
+
+**Note:** Version bump only for package @lion/button
+
+
+
+
+
 ## [0.3.31](https://github.com/ing-bank/lion/compare/@lion/button@0.3.30...@lion/button@0.3.31) (2019-10-25)
 
 **Note:** Version bump only for package @lion/button