+++ packages/calendar/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.11](https://github.com/ing-bank/lion/compare/@lion/calendar@0.2.10...@lion/calendar@0.2.11) (2019-11-06)
+
+**Note:** Version bump only for package @lion/calendar
+
+
+
+
+
+## [0.2.10](https://github.com/ing-bank/lion/compare/@lion/calendar@0.2.9...@lion/calendar@0.2.10) (2019-11-01)
+
+**Note:** Version bump only for package @lion/calendar
+
+
+
+
+
 ## [0.2.9](https://github.com/ing-bank/lion/compare/@lion/calendar@0.2.8...@lion/calendar@0.2.9) (2019-10-25)
 
 **Note:** Version bump only for package @lion/calendar