+++ packages/input-date/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.66](https://github.com/ing-bank/lion/compare/@lion/input-date@0.1.65...@lion/input-date@0.1.66) (2019-11-06)
+
+**Note:** Version bump only for package @lion/input-date
+
+
+
+
+
+## [0.1.65](https://github.com/ing-bank/lion/compare/@lion/input-date@0.1.64...@lion/input-date@0.1.65) (2019-11-01)
+
+**Note:** Version bump only for package @lion/input-date
+
+
+
+
+
 ## [0.1.64](https://github.com/ing-bank/lion/compare/@lion/input-date@0.1.63...@lion/input-date@0.1.64) (2019-10-25)
 
 **Note:** Version bump only for package @lion/input-date