+++ packages/popup/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.10](https://github.com/ing-bank/lion/compare/@lion/popup@0.3.9...@lion/popup@0.3.10) (2019-11-06)
+
+**Note:** Version bump only for package @lion/popup
+
+
+
+
+
+## [0.3.9](https://github.com/ing-bank/lion/compare/@lion/popup@0.3.8...@lion/popup@0.3.9) (2019-11-01)
+
+**Note:** Version bump only for package @lion/popup
+
+
+
+
+
 ## [0.3.8](https://github.com/ing-bank/lion/compare/@lion/popup@0.3.7...@lion/popup@0.3.8) (2019-10-25)
 
 **Note:** Version bump only for package @lion/popup