+++ packages/select-rich/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.13](https://github.com/ing-bank/lion/compare/@lion/select-rich@0.3.12...@lion/select-rich@0.3.13) (2019-11-06)
+
+**Note:** Version bump only for package @lion/select-rich
+
+
+
+
+
+## [0.3.12](https://github.com/ing-bank/lion/compare/@lion/select-rich@0.3.11...@lion/select-rich@0.3.12) (2019-11-01)
+
+**Note:** Version bump only for package @lion/select-rich
+
+
+
+
+
 ## [0.3.11](https://github.com/ing-bank/lion/compare/@lion/select-rich@0.3.10...@lion/select-rich@0.3.11) (2019-10-25)
 
 **Note:** Version bump only for package @lion/select-rich