+++ packages/select/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.63](https://github.com/ing-bank/lion/compare/@lion/select@0.1.62...@lion/select@0.1.63) (2019-11-06)
+
+**Note:** Version bump only for package @lion/select
+
+
+
+
+
+## [0.1.62](https://github.com/ing-bank/lion/compare/@lion/select@0.1.61...@lion/select@0.1.62) (2019-11-01)
+
+**Note:** Version bump only for package @lion/select
+
+
+
+
+
 ## [0.1.61](https://github.com/ing-bank/lion/compare/@lion/select@0.1.60...@lion/select@0.1.61) (2019-10-25)
 
 **Note:** Version bump only for package @lion/select