+++ packages/switch/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/ing-bank/lion/compare/@lion/switch@0.1.5...@lion/switch@0.1.6) (2019-11-06)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
+## [0.1.5](https://github.com/ing-bank/lion/compare/@lion/switch@0.1.4...@lion/switch@0.1.5) (2019-11-01)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
+## [0.1.4](https://github.com/ing-bank/lion/compare/@lion/switch@0.1.3...@lion/switch@0.1.4) (2019-10-30)
+
+**Note:** Version bump only for package @lion/switch
+
+
+
+
+
+## [0.1.3](https://github.com/ing-bank/lion/compare/@lion/switch@0.1.2...@lion/switch@0.1.3) (2019-10-28)
+
+
+### Bug Fixes
+
+* **switch:** cross-browser fixes ([cb747d9](https://github.com/ing-bank/lion/commit/cb747d9))
+
+
+
+
+
 ## [0.1.2](https://github.com/ing-bank/lion/compare/@lion/switch@0.1.1...@lion/switch@0.1.2) (2019-10-25)
 
 **Note:** Version bump only for package @lion/switch